Add new interface method which abstracts API details
The core of the change here, and to maintain backwards
compatibility I have added it as a new method.

* Method accepts a RavelinObject derived class
The advantage here is twofold,
a) it has already built in validation so user gets
that first hand
b) the object passed in can contain meta data on the
request path, so there is no need for the user to
supply it.

* Method accepts bools for backfill and score requests
This simplify client code as it does not need a
condition block to call a different method, but can
pass in some ivar or var passed though